+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# 0.13.1
+- refactor: return `ConnectionId` when establishing connection.
+
 # 0.13.0
 - feat: Add request-response functionality. [PR 339](https://github.com/dariusc93/rust-ipfs/pull/339)
 - refactor: provide topic in pubsub event. If a topic filter is supplied, the topic will be excluded from the event. [PR 337](https://github.com/dariusc93/rust-ipfs/pull/337)